requirements.txt
Django==5.0.4
python-dotenv==1.0.1
psycopg2==2.9.9
pillow==10.3.0
colorama==0.4.6
flake8==7.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
django-celery-beat==2.6.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
setuptools==69.5.1
requests==2.31.0
redis==5.0.4

